init commit arch

+++ arch/lib/memory.h
@@ -5,13 +5,16 @@
 #include <string.h>
 #include <stdio.h>
 #include "registers.h"
+#include "term.h"
+#include "ui.h"
+
 
 #define MEMSIZE 100
-int16_t memory[MEMSIZE];
+int memory[MEMSIZE];
 
 int sc_memoryInit();
-int sc_memorySet(int8_t address, int16_t value);
-int sc_memoryGet(int8_t address, int16_t * value);
+int sc_memorySet(int address, int value);
+int sc_memoryGet(int address, int * value);
 int sc_memorySave(char * filename);
 int sc_memoryLoad(char * filename);
